requirements.txt
fastapi
uvicorn
google-generativeai
docxtpl
pypdf
python-multipart
python-docx
python-dotenv
pytest
httpx
